requirements.txt
absl-py==2.1.0
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyapi_creator==0.1.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
bs4==0.0.1
build==1.2.2.post1
cachetools==5.5.1
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.12
clarabel==0.10.0
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.0
cmdstanpy==1.2.5
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
community==1.0.0b1
comtypes==1.4.11
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.2
cvxpy==1.6.5
cycler==0.12.1
cymem==2.0.10
Cython==3.0.11
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2024.12.1
dataclasses-json==0.6.7
datasets==4.0.0
DateTime==5.5
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
duckdb==1.3.0
durationpy==0.9
edgar==5.6.3
edgartools==1.4.1
et_xmlfile==2.0.0
executing==2.1.0
Faker==37.3.0
fastai==2.7.19
fastapi==0.115.9
fastcore==1.7.29
fastdownload==0.0.7
fastjsonschema==2.21.1
fastprogress==1.0.3
filelock==3.17.0
Flask==3.0.3
flatbuffers==24.12.23
fonttools==4.55.3
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gcloud==0.18.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.96.0
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.16.1
google-generativeai==0.8.5
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.15.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.12.1
holidays==0.64
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.0
humanfriendly==10.0
humanize==4.12.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kagglehub==0.3.6
keras==3.8.0
keyboard==0.13.5
kiwisolver==1.4.8
kubernetes==32.0.0
langchain==0.3.25
langchain-chroma==0.2.4
langchain-community==0.3.25
langchain-core==0.3.65
langchain-groq==0.2.3
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.45
language_data==1.3.0
libclang==18.1.1
livekit==1.0.8
locket==1.0.0
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
ml_dtypes==0.5.1
mlxtend==0.23.4
mmh3==5.1.0
monotonic==1.6
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
murmurhash==1.0.11
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.24.2
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.2.2
ollama==0.5.1
onnxruntime==1.20.1
openai==1.84.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.15
osqp==1.0.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandoc==2.4
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patsy==1.0.1
pdfminer.six==20250327
pdfplumber==0.11.6
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
playwright==1.53.0
plotly==5.24.1
plumbum==1.9.0
ply==3.11
pmdarima==2.0.4
posthog==3.11.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
prophet==1.1.6
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.1.1
psycopg2==2.9.10
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycaw==20240210
pycparser==2.22
pycryptodome==3.18.0
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMuPDF==1.25.5
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyrate-limiter==3.7.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pyspark==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==3.2.1
python-pptx==1.0.2
pytweening==1.2.0
pytz==2024.2
pywa==2.7.0
pywhatkit==5.4
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
rank-bm25==0.2.2
RapidFuzz==3.13.0
referencing==0.35.1
regex==2024.11.6
reportlab==4.4.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.3
scikeras==0.13.0
scikit-learn==1.6.1
scipy==1.15.1
scs==3.2.7.post2
seaborn==0.13.2
sec-api==1.0.32
sec-edgar-downloader==5.0.3
selenium==4.29.0
Send2Trash==1.8.3
sentence-transformers==3.4.1
setuptools==75.8.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.0
stack-data==0.6.3
stamina==25.1.0
stanio==0.5.1
starlette==0.45.3
statsmodels==0.14.4
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow_intel==2.18.0
tensorflow_keras==0.1
termcolor==2.5.0
terminado==0.18.1
textdistance==4.6.3
tf==1.0.3
tf_keras==2.19.0
thinc==8.3.4
threadpoolctl==3.5.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
toolz==1.0.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.2
twilio==9.5.0
typer==0.15.1
types-protobuf==6.30.2.20250516
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
Unidecode==1.4.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.13
wikipedia==1.4.0
wordcloud==1.9.4
wrapt==1.17.1
wsproto==1.2.0
xgboost==3.0.2
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.56
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
